+++ src/Support/Traits/DataContainerTrait.php
@@ -1,9 +1,19 @@
 <?php namespace Orchestra\Support\Traits;
 
 use Illuminate\Support\Arr;
+use Illuminate\Contracts\Encryption\Encrypter;
+use Illuminate\Contracts\Encryption\DecryptException;
+use Illuminate\Contracts\Encryption\EncryptException;
 
 trait DataContainerTrait
 {
+    /**
+     * The encrypter implementation.
+     *
+     * @var \Illuminate\Contracts\Encryption\Encrypter|null
+     */
+    protected $encrypter;
+
     /**
      * Item or collection.
      *
@@ -31,6 +41,29 @@ trait DataContainerTrait
     }
 
     /**
+     * Get an encrypted item value.
+     *
+     * @param  string  $key
+     * @param  mixed   $default
+     *
+     * @return mixed
+     */
+    public function secureGet($key, $default = null)
+    {
+        $value = $this->get($key, $default);
+
+        if ($this->encrypter instanceof Encrypter) {
+            try {
+                return $this->encrypter->decrypt($value);
+            } catch (DecryptException $e) {
+                //
+            }
+        }
+
+        return $value;
+    }
+
+    /**
      * Set a item value.
      *
      * @param  string  $key
@@ -41,6 +74,27 @@ trait DataContainerTrait
     public function set($key, $value = null)
     {
         return Arr::set($this->items, $key, value($value));
+    }
+
+    /**
+     * Set an ecrypted item value.
+     *
+     * @param  string  $key
+     * @param  mixed   $value
+     *
+     * @return mixed
+     */
+    public function secureSet($key, $value = null)
+    {
+        try {
+            if ($this->encrypter instanceof Encrypter) {
+                $value = $this->encrypter->encrypt($value);
+            }
+        } catch (EncryptException $e) {
+            //
+        }
+
+        return $this->set($key, $value);
     }
 
     /**